pyproject.toml
[tool.poetry]
name = "gesture-classification"
version = "0.1.0"
description = ""
authors = ["Sarosh Manzoor <sarosh.manzoor@outlook.com>"]
readme = "README.md"
packages = [{include = "gesture_classification"}]


[tool.poetry.dependencies]
python = "3.9.1"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
rarfile = "^4.1"
pandas = "1.5.0"
python-multipart = "^0.0.9"
cloudpickle = "^3.0.0"
streamlit = "^1.32.0"
tensorflow-macos = "2.14.0"
sktime = {extras = ["dl"], version = "^0.27.0"}
plotly = "^5.19.0"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
notebook = "^7.1.1"
matplotlib = "^3.8.3"
black = "^23.7.0"
seaborn = "^0.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gesture-classification = "gesture_classification.main:main"
